+++ src/tenants.ts
@@ -12,6 +12,7 @@ interface OneBlinkAppsTenant {
     olderIOSTime: Intl.DateTimeFormat
   }
   name: string
+  productShortName: string
 }
 type Locale = 'en-AU' | 'en-US'
 
@@ -64,6 +65,7 @@ const tenants = {
         'BE5wtYdaQW3z7DWc08rzNlOwPuituVWFRLW_lUMD78ZJatFOjio8fYDHaIpQCRyeKsJ5j4kLaFU374J4dM90iUc',
       intlFormats: generateFormatters('en-AU'),
       name: 'OneBlink',
+      productShortName: 'OneBlink LcS',
     },
     civicplus: {
       awsRegion: 'us-east-2',
@@ -73,6 +75,7 @@ const tenants = {
         'BLg2Dn9sYj1a0I3AcS22Fg71uubdMLwoemG8zfnPOljgFKB-5MR3FIxc2Mtt0AzM3zk2QWl3YzEy6EEwIUmz19k',
       intlFormats: generateFormatters('en-US'),
       name: 'CivicPlus',
+      productShortName: 'CivicOptimize',
     },
   },
   prod: {
@@ -84,6 +87,7 @@ const tenants = {
         'BADH0JnMngI0uFKUnbC79VGXy5d6WutccnEvVuFBMx--BrZtFAHGTgOBiABJXmE8_VHC92_jK5K-2qdP2kZeius',
       intlFormats: generateFormatters('en-AU'),
       name: 'OneBlink',
+      productShortName: 'OneBlink LcS',
     },
     civicplus: {
       awsRegion: 'us-east-2',
@@ -93,6 +97,7 @@ const tenants = {
         'BLoDtCutrC7tEd75x89zBaIyz3Fk8AeWOcABasV3YO4Tei5UO8WjJVPFyilNLYxeseaiKlgoa0DOh1HoR59M_G4',
       intlFormats: generateFormatters('en-US'),
       name: 'CivicPlus',
+      productShortName: 'CivicOptimize',
     },
   },
 }